+++ src/CurrencyConverter.php
@@ -13,10 +13,6 @@ use Brick\Math\RoundingMode;
 
 /**
  * Converts monies into different currencies, using an exchange rate provider.
- *
- * @todo Now that this class provides methods to convert to both Money and RationalMoney, it makes little sense to
- *       provide the context in the constructor, as this only applies to convert() and not convertToRational().
- *       This should probably be a parameter to convert().
  */
 final class CurrencyConverter
 {
@@ -26,19 +22,11 @@ final class CurrencyConverter
     private ExchangeRateProvider $exchangeRateProvider;
 
     /**
-     * The context of the monies created by this currency converter.
-     */
-    private Context $context;
-
-    /**
      * @param ExchangeRateProvider $exchangeRateProvider The exchange rate provider.
-     * @param Context|null         $context              A context to create the monies in, or null to use the default.
-     *                                                   The context only applies to convert(), not convertToRational().
      */
-    public function __construct(ExchangeRateProvider $exchangeRateProvider, ?Context $context = null)
+    public function __construct(ExchangeRateProvider $exchangeRateProvider)
     {
         $this->exchangeRateProvider = $exchangeRateProvider;
-        $this->context              = $context ?? new DefaultContext();
     }
 
     /**
@@ -48,6 +36,7 @@ final class CurrencyConverter
      *
      * @param MoneyContainer      $moneyContainer The Money, RationalMoney or MoneyBag to convert.
      * @param Currency|string|int $currency       The Currency instance, ISO currency code or ISO numeric currency code.
+     * @param Context|null        $context        A context to create the money in, or null to use the default.
      * @param int                 $roundingMode   The rounding mode, if necessary.
      *
      * @return Money
@@ -55,9 +44,11 @@ final class CurrencyConverter
      * @throws CurrencyConversionException If the exchange rate is not available.
      * @throws RoundingNecessaryException  If rounding is necessary and RoundingMode::UNNECESSARY is used.
      */
-    public function convert(MoneyContainer $moneyContainer, $currency, int $roundingMode = RoundingMode::UNNECESSARY) : Money
+    public function convert(MoneyContainer $moneyContainer, $currency, ?Context $context = null, int $roundingMode = RoundingMode::UNNECESSARY) : Money
     {
-        return $this->convertToRational($moneyContainer, $currency)->to($this->context, $roundingMode);
+        return $this
+            ->convertToRational($moneyContainer, $currency)
+            ->to($context ?? new DefaultContext(), $roundingMode);
     }
 
     /**