+++ tests/CurrencyConverterTest.php
@@ -50,7 +50,7 @@ class CurrencyConverterTest extends AbstractTestCase
             $this->expectException($expectedResult);
         }
 
-        $actualResult = $currencyConverter->convert($money, $toCurrency, $roundingMode);
+        $actualResult = $currencyConverter->convert($money, $toCurrency, null, $roundingMode);
 
         if (! $this->isExceptionClass($expectedResult)) {
             $this->assertMoneyIs($expectedResult, $actualResult);
@@ -97,8 +97,8 @@ class CurrencyConverterTest extends AbstractTestCase
             $moneyBag->add($money);
         }
 
-        $currencyConverter = new CurrencyConverter($exchangeRateProvider, $context);
-        $this->assertMoneyIs($total, $currencyConverter->convert($moneyBag, $currency, $roundingMode));
+        $currencyConverter = new CurrencyConverter($exchangeRateProvider);
+        $this->assertMoneyIs($total, $currencyConverter->convert($moneyBag, $currency, $context, $roundingMode));
     }
 
     public function providerConvertMoneyBag() : array
@@ -164,7 +164,7 @@ class CurrencyConverterTest extends AbstractTestCase
             $this->expectException($expectedResult);
         }
 
-        $actualResult = $currencyConverter->convert($rationalMoney, $toCurrency, $roundingMode);
+        $actualResult = $currencyConverter->convert($rationalMoney, $toCurrency, null, $roundingMode);
 
         if (! $this->isExceptionClass($expectedResult)) {
             $this->assertMoneyIs($expectedResult, $actualResult);